// src/api/generated/schemas/registerSubnetRequest.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * cid-api
 * LDAP 로그인 + JWT 인가 API
 * OpenAPI spec version: v1
 */

export interface RegisterSubnetRequest {
  /**
   * CIDR 주소
   * @minLength 0
   * @maxLength 30
   */
  subnetCidrAddr: string;
  /**
   * 서브넷 설명
   * @minLength 0
   * @maxLength 200
   */
  subnetDescp?: string;
  /**
   * VLAN ID
   * @minLength 0
   * @maxLength 20
   */
  vlanId?: string;
  /**
   * VRF 명
   * @minLength 0
   * @maxLength 50
   */
  vrfNm?: string;
  /** 상위 서브넷 ID (자기참조 FK) */
  upperSubnetId?: number;
  /** CI 고유 ID */
  ciId?: number;
}